Add JTAG configuration for ARM926 with XIO2001.
Boards like the cifX 50E have a XIO2001 PCIE to PCI bridge chip in the
scan chain together with the ARM926 CPU.

Please note that the adapter speed is set to 50kHz. This is very slow and
results in long transfer times. The hardware setup used to develop this
patch included an FTDI JTAG adapter, a "Papa Schlumpf" PCIE tester board,
a PCIE adapter board and finally a CIFX 50E. This is quite a long way for
a JTAG signal. In this setup 100kHz was already too fast.